requirements.txt
ccxt
mysql-connector-python
matplotlib
